fastapi_service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
safetensors
pillow
requests
